refactor(diff-viewer): terminal resize 配線を分離
Refs #224
